+++ intelmq/lib/pipeline.py
@@ -417,6 +417,9 @@ class Amqp(Pipeline):
         self.ssl = getattr(self.parameters,
                            "{}_pipeline_ssl".format(queues_type),
                            False)
+        self.exchange = getattr(self.parameters,
+                                "{}_pipeline_amqp_exchange".format(queues_type),
+                                "")
         self.load_balance_iterator = 0
         self.kwargs = {}
         if self.username and self.password:
@@ -486,7 +489,7 @@ class Amqp(Pipeline):
 
         retval = False
         try:
-            retval = self.channel.basic_publish(exchange='',
+            retval = self.channel.basic_publish(exchange=self.exchange,
                                                 routing_key=destination_queue,
                                                 body=message,
                                                 properties=self.properties,